requirements.txt
cryptography
django
djangorestframework
PyGithub
pyjwt
requests
travispy